requirements.txt
spidev
pyelftools
pyyaml
tqdm